Exclude all unbundled library files from domain substitution
Although only three are in the domain substitution
list, exempt all of them to be on the safe side.

Every url in these files is part of a comment, and
thus are safe to exclude from domain substitution.

This is prep work for making unbundling and domain
substitution independent of each other.