configure.sh: Fix hyphenation-related error
TEST is undefined at the point where we check
it's value, so move it down after where TEST
is set.

Also add the missing HYPHENATION test so that
HYPHENATION=0 doesn't error out.

While we're here, refactor slightly so that
we put both the HYPHENATION and LLVM_PGO_VER
code under the same TEST=0 gate.