Revert "Fix more missing header build errors"
After quite a bit of troubleshooting, it turned
out that this was a misconfiguration issue at my
end.

Unbundled brotli and libdrm somehow crept into my
setup, probably a copy paste error when running
the configuration script.

A complete reset and rebuild showed no errors
so the 'fix' patches can be ditched.

This reverts commit 27f8a5129aa8ec5532a7ee99883a19b514671790.