d/rules: Fix reset_unbundling
The problem here is that some libraries are
pulled in as DEPS submodules (via the gclient
sync command during build prep). These will
throw a non-existent file error whn attempting
to the git checkout command from the root
directory.

We can fix this by always changing into the
file's directory to perform the git checkout
command. This will work regardless of whether
the file in question is in-tree or part of a
submodule.