Enable partition_lock_priority_inheritance=true by default
This is supposed to be the futex equivalent to
the mutex priority inheritance option, and help
to increase overall responsiveness.

A patch is needed to fix a build error caused
by the scenario of mutex priority inheritance
being enabled when the partition lock variety
is disabled.

Add a PART_LOCK_PI configuration variable and
only enable the patch when PART_LOCK_PI=0 and
MUTEX_PI=1.

Also make it an error to set PART_LOCK_PI=1
when MUTEX_PI=0, as the former seems to be
tied in with the latter.

Added :-
optional/fixes/partition-lock-priority-inheritance.patch