node-get.sh: Download DEPS version of node
One good reason to get the 'official' chromium copy
of node (instead of the one from nodejs.org) is
that we then don't need to apply the node version
check patch, which is intended for system packaged
node.

The script grabs the node component's object_name
from the DEPS file to make the download url, to
ensure that the node version is up to date.