configure.sh: Drop LABS_TOOLBAR_BUTTON variable
The sole purpose of this variable is to enable
the '--chrome-labs' flag in /etc/chromium.d/ui.

We could have several separate variables to
enable/disable the many UI-related flags, but
this would clutter up the script for little
gain.

So let's drop this and leave it up to the user
to edit the ui flag file to their own taste.