configure.sh: Introduce BUILD_TS configuration variable
The blurb for the compute_build_timestamp build flag seems
to suggest that the timestamp is used during compilation.
It might be true that code that is unchanged apart from
containing a newer timestamp invokes recompilation instead
of the use of the cache.

We can patch build/timestamp.gni to using the non-official
build timestamp mechanism, which updates the timestamp
monthly. Or we can patch build/compute_build_timestamp.py
to produce a fixed timestamp.

The default for BUILD_TS is zero (disabled). A value of 1
will use the non-official build timestamp scheme. A value
of 2 uses a fixed timestamp (configurable via the TIMESTAMP
variable). When CCACHE=1, BUILD_TS is set to 1 by default.