configure.sh: Allow setting TIMESTAMP to imply BUILD_TS=2
There would be no point setting TIMESTAMP for any other
BUILD_TS value, so this allows setting TIMESTAMP without
specifying BUILD_TS.

Also shuffle the case statement sanitisation of TIMESTAMP
to near the top of the script. This allows us to exit
earlier in case the variable is set to an invalid value.

It's also more logical and neater to set TIMESTAMP before
we set BUILD_TS in the script.

Set TIMESTAMP=0 if it's initially unset, so that we can
have TIMESTAMP > 0 as a condition for editing the patch.