Add various/compiler-flags/fortify-level-three.patch
This raises the _FORTIFY_SOURCE level to 3 in accordance
with what Arch Linux is doing in their chromium package [1]
and a couple of Red Hat articles which recommend this [2-3].

We need to add '-Wno-macro-redefined' to our list of warning
suppression flags to get rid of a flood of useless console
messagaes.

Note that the raised fortify level only affects chromium code
plus exerts only a very minor performance penalty.

[1] https://gitlab.archlinux.org/archlinux/packaging/packages/chromium
[2] https://developers.redhat.com/articles/2022/09/17/gccs-new-fortification-level
[3] https://developers.redhat.com/articles/2023/02/06/how-improve-application-security-using-fortifysource3