d/control: Bump libicu-dev to 74.2-2
In spite of checking for CVEs several times since using
versions of libicu from the experimental branch, one was
published last month (CVE-2025-5222) [1-2] which has only
just come to my attention.

This means that using the old version from the debian
snapshot archive is no longer tenable from a security
standpoint.

I have made a separate repo [3] to host build files for
icu 74.2, which includes a patch from debian unstable to
fix the CVE. This is essentially the same patch as that
from stable-security (for 67.1), so should be equally
valid for 74.2. End users can download libicu74 from
the releases section of the same repo.

It's not an ideal solution, but if we unbundle icu then
we can unbundle harfbuzz-ng, libxslt and libxml too, so
it's worth doing.

Hopefully upstream chromium won't take too long to move
onto icu 76.

[1] https://www.cve.org/CVERecord?id=CVE-2025-5222
[2] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1106684
[3] https://github.com/berkley4/icu-74-debian