Go back to SYS_ICU=0 as our default
TLDR, going back to bundled icu will avert having
two different versions of icu system libraries
from being loaded, and get rid of the need to
install a third party libicu74 deb package.

Firstly, the interdependency between harfbuzz and
icu means that our harfbuzz system library depends
on libicu76 while UC itself depends on libicu74.

A bit of inconsistency in using different versions
of the same system library might have been a price
worth paying were it not for the fact that users
are having to install a third party libicu74 deb.

In mitigation, the third party deb (fixing a CVE)
was uploaded after the UC debs had already been
built (a lengthy process on a slow build machine).
It seemed the best way to fix what was a security
hole that was discovered post release.

This serves as a reminder to be careful when
venturing down the confusing rabbit hole of
unbundling dependencies.