configure.sh: Re-work setting/handling of DEPS_PATCH
Similar to RUST_INST, use a genarated list of
strings to correspond with which DEPS patches
have been applied.

This way we can use much simpler and clearer
case statements to gate things.

We also separate out the initialisation of
DEPS_PATCH from the intialisation of TEST to
further simplify and clarify what's being
done.

Should be no functional change.