configure.sh: Fix/improve handling of QT=6/QT_6=1
Move setting the default for QT_6 up the
script to where most other defaults are
set.

Imply QT=6 when QT_6=1, then we properly
have an alias and are able to enter the
any QT conditionals.

Add logic to error out if setting QT_6=1
together with QT=0, or setting QT=6 with
QT_6=0.

Fix the problem where QT_6=1 (or QT=6)
installed the Qt 5 verstion of the shim
instead of the QT 6 one.