configure.sh: Make GTK=0 and QT=0 an invalid choice
Pop up open file dialogues, eg to upload the deb files
on github won't work when both GTK and QT are disabled.
However, drag and drop works.

The real problem is that this configuration is probably
unpopular enough to not get tested. Clicking on the
'save and share/save webpage as' via the menu causes
a full crash.

We just need to revert the parts where both UI elements
are disables and keep the rest.