configure.sh: Drop TP_STORAGE_PART variable and fix flag
The TP_STORAGE_PART variable was used to control whether or
not to enable the runtime flag. This is another situation
where it's probably best left to the user to edit the flag
file directly [1].

As it happens, we were using an old flag, and the correct
flag can be seen from looking at the vanadinite patch [2].

Users who wanted to enable this feature were actually
running with it enabled by default anyway, as per the
patch disabling the origin trial (ie which enabled the
feature).

So let's use a commented out '--enable-features' flag
instead, to allow users to disable third party storage
partitioning.

We can also remove the FLAG_ISOL (non-interactive)
variable, since there's now only one sed command
operating on the flag file.

[1] /etc/chromium.d/isolation
[2] vanadinite/0012-enable-partitioning-third-party-storage.patch