widevine-cdm.sh: Fix/clean up/relocate the script
Fix an issue where CUR_DIR gets inadvertently set
to the script filename. Use the same technique as
in other scripts to enforce running the script from
its containing folder.

Move the widevine-cdm script folder to the build
directory so that it can share the download_cache
folder used by other scripts (it can still be run
outside of this location if the whole widevine-cdm
folder is moved elsewhere).