configure.sh: Allow customising shell launcher cpu check
This introduces the CPU configuration variable
to control the cpu instruction requirement in the
shell launcher.

The deb releases are built with avx enabled, but
builders configuring with AVX=0 (or MARCH=0) will
instead produce a shell launcher which enforces
an sse3 (or pni) requirement.

I don't think many builders would do this, but
those configuring with eg MARCH=x86-64-v3 could
set CPU to something like avx2 in order to allow
those builds to fail gracefully when run on
non-avx2 systems.