configure.sh: Allow producing shell launcher when TEST=1
The original reason given for not producing the shell
launcher in test mode was that an error would arise
when attempting to execute update_launcher.sh on a
partitions which was mounted noexec.

I'm going to drop this, as it's actually less hassle
to remount as executable than repeatedly re-editing
the script to remove the TEST=0 requirement.