configure.sh: Split AES_PCLMUL into AES and PCLMUL
The PCLMUL instruction set is the oldest and most
common instruction set that we enable, so it makes
sense to enable this first.

In terms of ubiquity, AES is always available on
anything newer than Skylake, and a fair bit earlier
on AMD processors.

AVX isn't universally present, at least in the case
of Pentium/Celeron, until Tiger Lake (2020). Similar
to AES, it generally features on AMD CPUs much
earlier.

In terms of patch order we now enable PCLMUL, AES
and AVX in order of ubiquity. Those lacking AES
will very likely lack AVX, so there shouldn't be
issues with patch context for most users.

Also, we previously set AES_PCLMUL=1 when BMI
and AVX were enabled. Don't replicate this with
the separate AES and PCLMUL variables, since
both are already enabled by default.