widevine-cdm.sh: Add a couple of aria2c options
The --conditional-get=true option should have
been added in the commit where remote timestamp
retrieval was added. This allows the download
only when the timestamp of the remote file is
newer than the local one.

Also add --allow-overwrite=true so that when a
newer file is downloaded, it clobbers the old
one instead of creating a backup copy of it.