configure.sh: Fix DNS_BUILTIN=1
To enable the built-in DNS client, we we not only
need to flip the policy key from false to true, but
also disable the kryptonite patch [1] which disables
the AsyncDns feature.

The patch might not even be needed when we're using
managed policy to disable, but ensures that the
feature gets properly disabled.

[1] kryptonite/disable-async-dns.patch