configure.sh: Fix non-default MARCH
This has been untested and neglected for far too long.

Firstly, we need to edit the correct patches (ie ones
which contain '-march=x86-64-v2' or '-mtune=generic')
only if MARCH is something other than 'x86-64-v2' or
MTUNE is not 'generic'.

Make greater use of the case statement where we catch
MARCH quirks, and use the fact that when arch_patches
is not null we have patches which need modification.

Do similar with rust-instructions.patch too, so that
we change target-cpu and only edit when needed.