shell launcher: Exit earlier if trying to debug with no gdb
If gdb isn't installed then this will avoid executing
a few commands by exiting as soon as the -g/--debug
switch is processed. This makes the final execution
section a bit clearer too.

Also use the 'command' shell builtin to test for the
existence of the gdb binary, as this should also pick
up gdb installed to locations other than /usr/bin.