Update handling of system rust when STABLE=1
On unstable, the recently updated rustc has a
rust related build error caused by application
of optional/system/rust.patch.

So only enable it on stable, which is now using
the same version that unstable was using.

We also don't need to enable it under upstream
rust (SYS_RUST=2) any more, since the stable
version of that is a much newer version than
that of debian unstable.

Also bump the rustc version requirement to
that of debian stable, and drop rustc-web as
an alternative dependency as it's no longer
available on debian stable.