Enable SYS_BROTLI by default for STABLE=1
This also enables freetype under stable, which
depends on brotli, as well as fontconfig (which
depends on freetype).

I'm going to keep the SYS_BROTLI stuff until I
can decide what to do with it.

We can drop the following patch, since the
newer freetype in stable won't need it :-

optional/system/freetype-COLRV1.patch