d/control: Make brotli versioned and always enabled
Nothing in the configuration script enabled the brotli
package dependency, but it's needed by the build system
at least when using the system brotli library.

It remains to be seen whether this is needed when using
bundled brotli, but there seems little harm in making
it a requirement. For stable builders it gives a nudge
to dist upgrade to the debian trixie release.