Remove the ability to disable zstd and drop SYS_ZSTD
Things started to break a couple of months back
when disabling zstd, so remove the ability to
disable it via the ZstdContentEncodingEnabled
managed policy key.

The disable_zstd_filter=true build flag in the
rules file is commented out and seems to be a
relic. This is removed along with the SYS_ZSTD
stuff.