configure.sh: Fix/re-jig CHROMECAST/MEDIA_REMOTING
Having MEDIA_REMOTING=1 and MEDIA_REMOTING_SET=0
was impossible, so we don't need to account for
this possibility (or advise users to explicity
set MEDIA_REMOTING=0).

Move the error condition under CHROMECAST=0 up
the script to where other things error out.

AFAIK, media remoting is only relevant to
chromecast, so move the whole MEDIA_REMOTING
section to be under the CHROMECAST one, and
default enable MEDIA_REMOTING in this case.