d/rules: Build with enable_on_device_translation=false
This seems to be a separate AI version of google
translate, with the actual translation done by the
browser instead of a remote server.

I'm not sure if it's intended to replace regular
google translate, but disable it for now.

Requires a patch to exclude sources and mask out
a couple of symbols.

Added :-

various/disable/on-device-translation.patch